requirements_api.txt
fastapi~=0.85.0
python-multipart~=0.0.5
psycopg2-binary~=2.9.5
sqlmodel~=0.0.8
uvicorn~=0.18.3
python-dotenv~=0.21.0
debugpy
aiogram
orjson
